split out answers into seperate file
